feat: extract event, category, compilation to event-service

+++ core/interaction-api/src/main/java/ru/yandex/practicum/dto/event/EventStatistics.java
@@ -3,7 +3,9 @@ package ru.yandex.practicum.dto.event;
 import java.util.Map;
 
 public record EventStatistics(
-        Map<Long, Integer> confirmedRequests,
+
+        Map<Long, Long> confirmedRequests,
+
         Map<Long, Long> views
 ) {
 }
